Lisa/requirements.txt
ace_tools==0.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.12
defusedxml==0.7.1
distro==1.9.0
dotenv==0.9.9
Flask==3.1.0
flask-cors==5.0.1
fonttools==4.56.0
fpdf2==2.8.2
gitdb==4.0.12
GitPython==3.1.44
gmpy2==2.2.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
llvmlite==0.44.0
load-dotenv==0.1.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
narwhals==1.32.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai==1.69.0
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.4
PubChemPy==1.0.4
pyadic==0.2.3
pyarrow==19.0.1
pycryptodome==3.22.0
pydantic==2.11.1
pydantic_core==2.33.0
pydeck==0.9.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.24.0
sagemath==1.3.0
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.44.0
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.6.0
tiktoken==0.9.0
tk==0.1.0
toml==0.10.2
tornado==6.4.2
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
tzdata==2025.2
urllib3==2.3.0
waitress==3.0.2
watchdog==6.0.0
Werkzeug==3.1.3
